modify addToRam to use ram array indices

+++ include/pcb.h
@@ -1,6 +1,6 @@
 #ifndef PCB_H
 #define PCB_H
     #include <stdint.h>
-    typedef struct {uintptr_t PC; uintptr_t start; uintptr_t end;} PCB;
-    PCB* makePCB(uintptr_t start, uintptr_t end);
+    typedef struct {int PC; int start; int end;} PCB;
+    PCB* makePCB(int start, int end);
 #endif
