+++ include/ram.h
@@ -2,6 +2,6 @@
 #define RAM_H
     #include <stdio.h>
     #include <stdint.h>
-    void addToRAM(FILE *p, uintptr_t *start, uintptr_t *end);
+    void addToRAM(FILE *p, int *start, int *end);
     #define RAM_SIZE 1000
 #endif
